requirements.txt
# AutoHelp.uz Telegram Bot - Dependencies
# Core
aiogram>=3.15.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.1

# Redis
redis[hiredis]>=5.2.1

# Web Panel
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
jinja2>=3.1.5
python-multipart>=0.0.20

# Background Tasks
apscheduler>=3.10.4

# Excel Reports
openpyxl>=3.1.5

# Configuration
pydantic-settings>=2.7.1
python-dotenv>=1.0.1

# Logging
loguru>=0.7.3

# Utilities
aiofiles>=24.1.0
babel>=2.16.0
orjson>=3.10.14
